requirements.txt
streamlit
pandas
numpy
python-binance
TA-Lib
pandas-ta
scikit-learn
imbalanced-learn
lightgbm
plotly
